Add files via upload
The Phonebook Application is a simple yet effective program designed for managing personal contacts. Developed using Python and SQLite database, this project was created by our group using PyCharm as the IDE. The application allows users to perform essential operations such as adding, editing, and deleting contacts. It can be further enhanced with additional features, such as storing more detailed information or exporting contacts to a file for backup and sharing.